[EPT-1709_SEPARATE_ACTIONS_TO_NEW_FILE] Updating edit and delete actions to new file
