+++ src/ListView/PageTable.jsx
@@ -1,6 +1,7 @@
 import React from "react";
+import { Table } from "@sailthru/stui-components";
+import PageTableActions from "./PageTableActions";
 
-import { Table, HoverDisclosure, OptionsMenu } from "@sailthru/stui-components";
 
 /** Configures STUI Table to render data in columns */
 function PageTable({ loading, pages, getPageById, displayModal }) {
@@ -84,33 +85,7 @@ function PageTable({ loading, pages, getPageById, displayModal }) {
       accessor: (p) => p.pageId,
       width: 0.2,
       disableSortBy: true,
-      Cell: function ({ cell }) {
-        const pageId = cell.value;
-        const clickOptions = [
-          {
-            label: "Edit",
-            icon: "fal fa-edit fa-fw",
-            iconHover: "fas fa-edit fa-fw",
-            onClick: () => window.open(`/page?page_id=${pageId}`, "_self"), // TODO see EPT-1713 (add routing?)
-          },
-          {
-            label: "Delete",
-            icon: "fal fa-trash fa-fw",
-            iconHover: "fal fa-trash fa-fw",
-            onClick: () => {
-              console.log("Deleting", pageId);
-              displayModal({
-                // TODO see EPT-1713
-              });
-            },
-          },
-        ];
-        return (
-          <HoverDisclosure disclosureWidth="150px">
-            <OptionsMenu options={clickOptions} />
-          </HoverDisclosure>
-        );
-      },
+      Cell: (cell) => (<PageTableActions pageId={cell.value} />)
     },
   ];
